fix(runner): normalize scalar tags across runners
