+++ packages/bruno-common/src/index.ts
@@ -1,7 +1,7 @@
 export { mockDataFunctions, timeBasedDynamicVars } from './utils/faker-functions';
 export { default as interpolate, interpolateObject } from './interpolate';
 export { percentageToZoomLevel } from './zoom';
-export { default as isRequestTagsIncluded } from './tags';
+export { default as isRequestTagsIncluded, normalizeTags } from './tags';
 export { transformExampleStatusInCollection } from './example-status';
 
 export * as utils from './utils';
